Reduce dependency over Flags class.  This will be deprecated so that adhoc flags can be added without any additional manual configuration
