+++ proxy/common/constants.py
@@ -86,3 +86,7 @@ PLUGIN_PAC_FILE = 'proxy.http.server.HttpWebServerPacFilePlugin'
 PLUGIN_DEVTOOLS_PROTOCOL = 'proxy.http.inspector.DevtoolsProtocolPlugin'
 PLUGIN_DASHBOARD = 'proxy.dashboard.dashboard.ProxyDashboard'
 PLUGIN_INSPECT_TRAFFIC = 'proxy.dashboard.inspect_traffic.InspectTrafficPlugin'
+
+PY2_DEPRECATION_MESSAGE = '''DEPRECATION: proxy.py no longer supports Python 2.7.  Kindly upgrade to Python 3+. '
+                'If for some reasons you cannot upgrade, use'
+                '"pip install proxy.py==0.3".'''
