+++ tests/http/test_web_server.py
@@ -15,6 +15,7 @@ import unittest
 import selectors
 from unittest import mock
 
+from proxy.proxy import Proxy
 from proxy.common.flags import Flags
 from proxy.core.connection import TcpClientConnection
 from proxy.http.handler import HttpProtocolHandler
@@ -34,7 +35,7 @@ class TestWebServerPlugin(unittest.TestCase):
         self._conn = mock_fromfd.return_value
         self.mock_selector = mock_selector
         self.flags = Flags()
-        self.flags.plugins = Flags.load_plugins([
+        self.flags.plugins = Proxy.load_plugins([
             b'proxy.http.proxy.HttpProxyPlugin',
             b'proxy.http.server.HttpWebServerPlugin',
         ])
@@ -97,7 +98,7 @@ class TestWebServerPlugin(unittest.TestCase):
                 events=selectors.EVENT_READ,
                 data=None), selectors.EVENT_READ), ]
         flags = Flags()
-        flags.plugins = Flags.load_plugins([
+        flags.plugins = Proxy.load_plugins([
             b'proxy.http.proxy.HttpProxyPlugin',
             b'proxy.http.server.HttpWebServerPlugin',
         ])
@@ -150,7 +151,7 @@ class TestWebServerPlugin(unittest.TestCase):
         flags = Flags(
             enable_static_server=True,
             static_server_dir=static_server_dir)
-        flags.plugins = Flags.load_plugins([
+        flags.plugins = Proxy.load_plugins([
             b'proxy.http.proxy.HttpProxyPlugin',
             b'proxy.http.server.HttpWebServerPlugin',
         ])
@@ -200,7 +201,7 @@ class TestWebServerPlugin(unittest.TestCase):
                 data=None), selectors.EVENT_WRITE)], ]
 
         flags = Flags(enable_static_server=True)
-        flags.plugins = Flags.load_plugins([
+        flags.plugins = Proxy.load_plugins([
             b'proxy.http.proxy.HttpProxyPlugin',
             b'proxy.http.server.HttpWebServerPlugin',
         ])
@@ -238,7 +239,7 @@ class TestWebServerPlugin(unittest.TestCase):
 
     def init_and_make_pac_file_request(self, pac_file: str) -> None:
         flags = Flags(pac_file=pac_file)
-        flags.plugins = Flags.load_plugins([
+        flags.plugins = Proxy.load_plugins([
             b'proxy.http.proxy.HttpProxyPlugin',
             b'proxy.http.server.HttpWebServerPlugin',
             b'proxy.http.server.HttpWebServerPacFilePlugin',
